backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
psycopg2-binary==2.9.9
requests==2.32.3
python-dotenv==1.0.1
boto3==1.35.36
PyJWT==2.9.0